requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
clickhouse-driver==0.2.6
pandas==2.2.0
numpy==1.26.4
python-binance==1.0.19
dash==2.14.2
plotly==5.18.0
confluent-kafka==2.3.0
